set up remaning metadata
